ref/requirements.txt
# Dune Query Executor Dependencies
requests>=2.31.0
python-dotenv>=1.0.0